CHG move list_repos_with_latest_commit and async cache results in reaper
